+++ CHANGELOG.md
@@ -18,7 +18,7 @@ See DataLab [roadmap page](https://datalab-platform.com/en/contributing/roadmap.
     * Partial implementation of [Issue #216](https://github.com/DataLab-Platform/DataLab/issues/216).
     * Default strategy is now "Next power of 2".
 
-* New bilinear image generator.
+* New 2D ramp image generator.
   * This closes [Issue #203](https://github.com/DataLab-Platform/DataLab/issues/203).
 
 * New I/O features: